backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
psycopg[binary]==3.3.3
pydantic-settings==2.13.1
python-multipart==0.0.22
PyJWT==2.12.1
httpx==0.28.1
markdown-it-py==4.0.0
bleach==6.3.0
pytest==9.0.2
